# moonprint/bin/src/yaml/moonfobl_swagger-2.0.yaml
swagger: '2.0'
info:
  title: MOOn API
  description: API per il backend del sistema MOOn di modulistica online.
  version: 1.0.0
host: 'mydevpc.csi.it:10110'
schemes:
  - http
basePath: /restfacade/be
produces:
  - application/json
paths:
...
  /istanze:
    get:
      summary: Recurepa i dati principali di compilazioni utente  (filtrable con parametro "query")
      operationId: getIstanze
      tags:
        - business
      produces:
        - application/json
      parameters:
        - in: query
          name: fields
          type: string
          required: false
      responses:
        '200':
          description: lista dei dati principali di compilazioni utente
          schema:
            type: array
            items:
              $ref: '#/definitions/Istanza'
        default:
          description: errore generico
          schema:
            $ref: '#/definitions/Error'
    post:
      summary: Salva i dati di un form compilati dall'utente (inserisci o aggiorna secondo presenza idIstanza)
      operationId: saveIstanza
      tags:
        - business
      produces:
        - application/json
      consumes:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/Istanza'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Istanza'
  /istanze/{idIstanza}:
    get:
      summary: Recurepa i dati completi di una compilazione utente
      operationId: getIstanzaById
      tags:
        - business
      produces:
        - application/json
      parameters:
        - in: path
          name: idIstanza
          type: integer
          format: int64
          required: true
      responses:
        '200':
          description: dati di una compilazione utente di un modulo
          schema:
            $ref: '#/definitions/Istanza'
        default:
          description: errore generico
          schema:
            $ref: '#/definitions/Error'
    put:
      summary: Aggiorna i dati di un form compilati dall'utente
      operationId: putIstanza
      tags:
        - business
      produces:
        - application/json
      consumes:
        - application/json
      parameters:
        - in: path
          name: idIstanza
          type: integer
          format: int64
          required: true
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/Istanza'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Istanza'
...
definitions:
  Istanza:
    type: object
    properties:
      idIstanza:
        type: integer
        format: int64
        description: l'identificativo dell'istanza (Long)
      codiceIstanza:
        type: string
        description: l'identificativo utente dell'istanza secondo progressivo modulo
      stato:
        $ref: '#/definitions/Stato'
      created:
        type: string
        format: date-time
      modified:
        type: string
        format: date-time
      data:
        description: i dati del modulo compilati dall'utente
        type: object
      modulo:
        $ref: '#/definitions/Modulo'
      owner:
        type: string
      metadata:
        type: object
  Stato:
    type: object
    properties:
      idStato:
        type: integer
        description: l'identificativo dello stato
      nome:
        type: string
        description: il nome dello stato ('bozza','inviato')
      descrizione:
        type: string
        description: la descrizione dello stato
...
  Modulo:
    type: object
    properties:
      idModulo:
        type: integer
        format: int64
        description: l'identificativo del modulo (Long)
      codiceModulo:
        type: string
        description: il codice del modulo
      versioneModulo:
        type: string
        description: la versione del modulo
      oggettoModulo:
        type: string
        description: l'oggetto del modulo
      descrizioneModulo:
        type: string
        description: la descrizione del modulo
      dataIns:
        type: string
        format: date-time
        description: la data/oraio di inserimento/creazione del modulo
      dataUpd:
        type: string
        format: date-time
        description: la data/orario di ultimo aggiornamento del modulo
      flagIsRiservato:
        type: boolean
        description: true se il modulo e' riservato
      flagUsaProgressivo:
        type: boolean
        description: true se il modulo utilizzo il progressivo
      flagProtocolloIntegrato:
        type: boolean
        description: true se deve essere protocollato
      attoreUpd:
        type: string
        description: l'attore di inserimento/creazione o ultimo aggiornamento del modulo
      idModuloStruttura:
        type: integer
        format: int64
        description: l'identificativo del modulo struttura (Long)
      tipoStruttura:
        type: string
        description: il tipo della struttura WIZ-Wizard FRM-Form
      struttura:
        type: string
        description: la struttura del form in formato JSON formIO
...
  Error:
    type: object
    properties:
      code:
        type: string
      messaggioCittadino:
        type: string
      errorMessage:
        type: string
      fields:
        type: string
        x-nullable: true